futureroot/requirements.txt
asgiref==3.9.1
autopep8==2.3.2
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
idna==3.10
psycopg2==2.9.10
pycodestyle==2.14.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
